requirements.txt
# requirements.txt

# Core Data Stack

pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0


# NLP / Text Pipeline

nltk>=3.8.1
vaderSentiment>=3.3.2
sentence-transformers>=2.2.2
regex>=2023.0.0


# Machine Learning

scikit-learn>=1.3.0
joblib>=1.3.0


# Config + Pipeline Ops

pyyaml>=6.0
tqdm>=4.66.0


# Visualization

matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0


# Dashboard

streamlit>=1.28.0


# Model Interpretability

shap>=0.44.0


# MLOps Ready

mlflow>=2.8.0
rich>=13.0.0